+++ tests/testthat/test-run_mod.R
@@ -28,7 +28,7 @@ test_that(
     
     results |>
       attributes() |>
-      rlist::list.remove(c("row.names", "fitted_residuals")) |>
+      rlist::list.remove(c("row.names", "fitted_residuals", "population_params")) |>
       expect_snapshot_value(style = "deparse")
     
     results |>
@@ -68,7 +68,7 @@ test_that(
     
     results |>
       attributes() |>
-      rlist::list.remove(c("row.names", "fitted_residuals")) |>
+      rlist::list.remove(c("row.names", "fitted_residuals", "population_params")) |>
       expect_snapshot_value(style = "deparse")
     
     results |>
@@ -107,7 +107,7 @@ test_that(
     
     results |>
       attributes() |>
-      rlist::list.remove(c("row.names", "fitted_residuals")) |>
+      rlist::list.remove(c("row.names", "fitted_residuals", "population_params")) |>
       expect_snapshot_value(style = "deparse")
     
     results |>
@@ -147,7 +147,7 @@ test_that(
     
     results |>
       attributes() |>
-      rlist::list.remove(c("row.names", "jags.post", "fitted_residuals")) |>
+      rlist::list.remove(c("row.names", "jags.post", "fitted_residuals", "population_params")) |>
       expect_snapshot_value(style = "serialize")
     
     results |>
@@ -185,7 +185,7 @@ test_that(
     
     results |>
       attributes() |>
-      rlist::list.remove(c("row.names", "fitted_residuals")) |>
+      rlist::list.remove(c("row.names", "fitted_residuals", "population_params")) |>
       expect_snapshot_value(style = "serialize")
     
     results |>